refactor: categorized files
