+++ app/src/main/java/inc/padmal/ussdtrack/Controllers/SMSReceiver.java
@@ -1,12 +1,12 @@
-package inc.padmal.ussdtrack;
+package inc.padmal.ussdtrack.Controllers;
 
-import android.app.Service;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
 import android.telephony.SmsMessage;
-import android.util.Log;
+
+import inc.padmal.ussdtrack.Controllers.SmsListener;
 
 /**
  * Created by knight on 10/21/18.
