+++ app/src/main/java/inc/padmal/ussdtrack/Controllers/SmsListener.java
@@ -1,4 +1,4 @@
-package inc.padmal.ussdtrack;
+package inc.padmal.ussdtrack.Controllers;
 
 /**
  * Created by knight on 10/21/18.
