+++ app/src/main/java/inc/padmal/ussdtrack/Models/Medium.java
@@ -1,4 +1,4 @@
-package inc.padmal.ussdtrack;
+package inc.padmal.ussdtrack.Models;
 
 import android.content.Context;
