+++ app/src/main/java/inc/padmal/ussdtrack/TrackingActivity.java
@@ -32,6 +32,11 @@ import com.github.mikephil.charting.interfaces.datasets.ILineDataSet;
 import org.json.JSONException;
 import org.json.JSONObject;
 
+import inc.padmal.ussdtrack.Controllers.SMSReceiver;
+import inc.padmal.ussdtrack.Controllers.SmsListener;
+import inc.padmal.ussdtrack.Helpers.Plotter;
+import inc.padmal.ussdtrack.Helpers.USSDService;
+
 public class TrackingActivity extends AppCompatActivity
         implements NavigationView.OnNavigationItemSelectedListener {
